Refactor required documents information into its own folder that is consumed by a container handler
