+++ client/presentations/documents/index.js
@@ -1,0 +1,17 @@
+'use strict';
+
+import CaliforniaResidencyDocuments  from './california-residency-documents.jsx';
+import MedicalDocuments              from './medical-documents.jsx';
+import RealIdDocuments               from './real-id-documents.jsx';
+import SocialSecurityDocuments       from './social-security-documents.jsx';
+import VeteransDocuments             from './veterans-documents.jsx';
+import BulletList                    from './bullet-list.jsx';
+
+export {
+  CaliforniaResidencyDocuments,
+  MedicalDocuments,
+  RealIdDocuments,
+  SocialSecurityDocuments,
+  VeteransDocuments,
+  BulletList
+};
